+++ src/main/java/com/example/samplerest/customer/CustomerRepository.java
@@ -2,12 +2,12 @@ package com.example.samplerest.customer;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigDecimal;
 import java.util.Optional;
 
-public interface CustomerRepository extends JpaRepository<Customer, BigDecimal> {
+public interface CustomerRepository extends JpaRepository<Customer, Long> {
 
     Optional<Customer> findCustomerByEmail(String email);
-    Customer findCustomerById(BigDecimal id);
+
+    Customer findCustomerById(Long id);
 
 }
